refactor: mv src/lib/gui/TrayIcon => src/gui/src/TrayIcon
