+++ src/gui/src/TrayIcon.h
